+++ lib/sleeping_king_studios/tools/toolbox/constant_map.rb
@@ -1,4 +1,4 @@
-# lib/sleeping_king_studios/tools/toolbox/constant_map.rb
+# frozen_string_literal: true
 
 require 'sleeping_king_studios/tools/string_tools'
 require 'sleeping_king_studios/tools/toolbox'
@@ -10,17 +10,17 @@ module SleepingKingStudios::Tools::Toolbox
       # Creates a new ConstantMap.
       #
       # @param constants [Hash] The constants to define.
-      def new constants
+      def new(constants)
         mod = Module.new
         mod.extend self
 
         constants.each do |const_name, const_value|
           mod.const_set const_name, const_value
-        end # each
+        end
 
         mod
-      end # class method new
-    end # eigenclass
+      end
+    end
 
     # Returns a hash with the names and values of the defined constants.
     #
@@ -28,17 +28,17 @@ module SleepingKingStudios::Tools::Toolbox
     def all
       constants.each.with_object({}) do |const_name, hsh|
         hsh[const_name] = const_get(const_name)
-      end # each
-    end # method all
+      end
+    end
 
     # Iterates through the defined constants, yielding the name and value of
     # each constant to the block.
     #
     # @yieldparam key [Symbol] The name of the symbol.
     # @yieldparam value [Object] The value of the symbol.
-    def each &block
+    def each(&block)
       all.each(&block)
-    end # method each
+    end
 
     # Freezes the constant map and recursively freezes every constant value
     # using ObjectTools#deep_freeze. Also pre-emptively defines any reader
@@ -49,46 +49,48 @@ module SleepingKingStudios::Tools::Toolbox
       constants.each do |const_name|
         reader_name = const_name.downcase
 
-        define_reader(const_name, reader_name) unless methods.include?(reader_name)
+        unless methods.include?(reader_name)
+          define_reader(const_name, reader_name)
+        end
 
         object_tools.deep_freeze const_get(const_name)
-      end # each
+      end
 
       super
-    end # method freeze
+    end
 
     private
 
-    def define_reader const_name, reader_name = nil
+    def define_reader(const_name, reader_name = nil)
       reader_name ||= const_name.downcase
 
-      define_singleton_method reader_name, ->() { const_get const_name }
-    end # method define_reader
+      define_singleton_method reader_name, -> { const_get const_name }
+    end
 
-    def method_missing symbol, *args, &block
+    def method_missing(symbol, *args, &block)
       const_name = string_tools.underscore(symbol.to_s).upcase.intern
 
       if constants.include?(const_name)
         define_reader(const_name, symbol)
 
         return send(symbol, *args, &block)
-      end # if
+      end
 
       super
-    end # method method_missing
+    end
 
     def object_tools
       ::SleepingKingStudios::Tools::ObjectTools
-    end # method object_tools
+    end
 
-    def respond_to_missing? symbol, include_all = false
+    def respond_to_missing?(symbol, include_all = false)
       const_name = string_tools.underscore(symbol.to_s).upcase.intern
 
       constants.include?(const_name) || super
-    end # method respond_to_missing?
+    end
 
     def string_tools
       ::SleepingKingStudios::Tools::StringTools
-    end # method string_tools
-  end # module
-end # module
+    end
+  end
+end