+++ lib/sleeping_king_studios/tools/toolbox/mixin.rb
@@ -1,4 +1,4 @@
-# lib/sleeping_king_studios/tools/toolbox/mixin.rb
+# frozen_string_literal: true
 
 require 'sleeping_king_studios/tools/toolbox'
 
@@ -6,27 +6,27 @@ module SleepingKingStudios::Tools::Toolbox
   # Implements recursive inheritance of both class and instance methods.
   module Mixin
     # @api private
-    def self.mixin? mod
+    def self.mixin?(mod)
       return false unless mod.is_a?(Module)
 
       mod.singleton_class.include?(self)
-    end # class method mixin?
+    end
 
     # @api private
-    def included other
+    def included(other)
       return super unless defined?(self::ClassMethods)
 
       if SleepingKingStudios::Tools::Toolbox::Mixin.mixin?(other)
         unless other.constants(false).include?(:ClassMethods)
           other.const_set(:ClassMethods, Module.new)
-        end # unless
+        end
 
-        other::ClassMethods.send :include, self::ClassMethods
+        other::ClassMethods.include(self::ClassMethods)
       else
         other.extend self::ClassMethods
-      end # if-else
+      end
 
       super
-    end # method included
-  end # module
-end # module
+    end
+  end
+end